+++ src/layouts/default.vue
@@ -512,4 +512,19 @@ img {
         display: none !important;
     }
 }
+
+.__partner-info {
+    display: flex;
+    gap: 8px;
+    align-items: center;
+    justify-content: center;
+
+    font-size: 14px;
+
+    img {
+        width: 56px;
+        min-width: 56px;
+        max-width: 56px;
+    }
+}
 </style>