+++ src/server/plugins/vatsim.ts
@@ -1,5 +1,4 @@
 import type { VatsimDivision, VatsimEvent, VatsimSubDivision } from '~/types/data/vatsim';
-import type { Sigmets } from '~/utils/backend/storage';
 import { radarStorage } from '~/utils/backend/storage';
 import { updateAirlines, updateAustraliaData, updateTransceivers } from '~/utils/backend/vatsim/update';
 import { getRedis } from '~/utils/backend/redis';
@@ -41,7 +40,4 @@ export default defineNitroPlugin(app => {
     defineCronJob('* * * * * *', updateTransceivers);
     defineCronJob('15 * * * *', updateAustraliaData);
     defineCronJob('15 0 * * *', updateAirlines);
-    defineCronJob('15 * * * *', async () => {
-        radarStorage.sigmets = await $fetch<Sigmets>('https://aviationweather.gov/api/data/isigmet?format=geojson');
-    });
 });